percentage() -  allow percent>UINT8_MAX

+++ speeduino/maths.h
@@ -257,7 +257,7 @@ static inline uint32_t percentageApprox(uint8_t percent, uint32_t value) {
  * @param percent The percent to apply to value
  * @param value The value to operate on
  */
-static inline uint32_t percentage(uint8_t percent, uint32_t value) 
+static inline uint32_t percentage(uint16_t percent, uint32_t value) 
 {
     return (uint32_t)div100((uint32_t)value * (uint32_t)percent);
 }
